Improve performance of saving sessions.
The overhead of fsync_all() is too significant for saving sessions, therefore it is
removed, trading performance for safety in case of OS crashes or hardware failures.

Furthermore remove the random suffix for temporary files, which does not provide
much benefit (save operations on a specific file must be serialised on a higher
level).

Includes a first simple benchmark for saving sessions.